Add upiv1 implementation in PyFuncModel
